+++ src/main/java/uk/gov/hmcts/reform/emclient/application/EvidenceManagementClientApplication.java
@@ -14,8 +14,8 @@ import org.springframework.retry.annotation.EnableRetry;
 import uk.gov.hmcts.reform.authorisation.ServiceAuthorisationApi;
 import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGenerator;
 import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGeneratorFactory;
-import uk.gov.hmcts.reform.emclient.idam.api.IdamApiClient;
 import uk.gov.hmcts.reform.authorisation.healthcheck.ServiceAuthHealthIndicator;
+import uk.gov.hmcts.reform.divorce.idam.api.IdamApiClient;
 
 @SpringBootApplication
 @ComponentScan(basePackages = "uk.gov.hmcts", excludeFilters = {
