--- src/main/java/uk/gov/hmcts/reform/emclient/idam/models/UserDetails.java
@@ -1,20 +1,0 @@
-package uk.gov.hmcts.reform.emclient.idam.models;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.Builder;
-import lombok.Getter;
-
-import java.util.List;
-
-@Getter
-@Builder
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class UserDetails {
-
-    private String id;
-    private String email;
-    private String forename;
-    private String surname;
-    private List<String> roles;
-
-}
