+++ src/test/java/uk/gov/hmcts/reform/emclient/service/EvidenceManagementDeleteServiceImplTest.java
@@ -17,8 +17,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.client.ResourceAccessException;
 import org.springframework.web.client.RestTemplate;
 import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGenerator;
-import uk.gov.hmcts.reform.emclient.idam.models.UserDetails;
-import uk.gov.hmcts.reform.emclient.idam.services.UserService;
+import uk.gov.hmcts.reform.divorce.idam.models.UserDetails;
+import uk.gov.hmcts.reform.divorce.idam.services.UserService;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
